requirements.txt
Flask
gunicorn
google-generativeai
Flask-Cors
python-dotenv
easyocr
torch
torchvision